# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: Greater Bayelsa leadership development platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /voters:
    get:
      operationId: listVoters
      tags: [voters]
      summary: List voter roll with optional filters (HQ)
      parameters:
        - name: search
          in: query
          required: false
          schema: { type: string }
        - name: villageId
          in: query
          required: false
          schema: { type: integer }
      responses:
        "200":
          description: Voters
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Voter"
  /voters/import:
    post:
      operationId: importVoters
      tags: [voters]
      summary: Bulk import voter roll rows (HQ)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VoterImportBody"
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoterImportResult"

  # ---------- Me ----------
...
  /trainings:
    get:
      tags: [training]
      operationId: listTrainings
      summary: List training sessions visible to the caller (with own registration state)
      responses:
        "200":
          description: Training sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TrainingSession"
    post:
      tags: [training]
      operationId: createTraining
      summary: Schedule a training session (HQ only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrainingSessionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrainingSession"
  /trainings/{id}/register:
    post:
      tags: [training]
      operationId: registerTraining
      summary: Register interest/attendance for a training session (member)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrainingRegistration"
        "409":
          description: Already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /trainings/{id}/progress:
    post:
      tags: [training]
      operationId: updateTrainingProgress
      summary: Mark attendance/completion for registrants (HQ or session-village secretary)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrainingProgressBody"
      responses:
        "200":
          description: Updated registrations
          content:
            application/json:
              schema:
                type: object
                required: [updated]
                properties:
                  updated: { type: integer }
  /trainings/{id}/registrants:
    get:
      tags: [training]
      operationId: listTrainingRegistrants
      summary: List registrants for a session (HQ or village executive)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Registrants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TrainingRegistration"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string }
...
    Voter:
      type: object
      required: [id, firstName, lastName]
      properties:
        id: { type: integer }
        firstName: { type: string }
        lastName: { type: string }
        gender: { type: [string, "null"] }
        phone: { type: [string, "null"] }
        vin: { type: [string, "null"] }
        dateOfBirth: { type: [string, "null"] }
        occupation: { type: [string, "null"] }
        villageId: { type: [integer, "null"] }
        unitId: { type: [integer, "null"] }
        pollingUnit: { type: [string, "null"] }
    VoterImportBody:
      type: object
      required: [rows]
      properties:
        rows:
          type: array
          minItems: 1
          maxItems: 5000
          items:
            type: object
            required: [firstName, lastName]
            properties:
              firstName: { type: string }
              lastName: { type: string }
              gender: { type: [string, "null"] }
              phone: { type: [string, "null"] }
              vin: { type: [string, "null"] }
              dateOfBirth: { type: [string, "null"] }
              occupation: { type: [string, "null"] }
              villageName: { type: [string, "null"] }
              unitName: { type: [string, "null"] }
    VoterImportResult:
      type: object
      required: [imported, skipped]
      properties:
        imported: { type: integer }
        skipped: { type: integer }
...
    TrainingSession:
      type: object
      required: [id, title, scheduledOn, status, registeredCount, createdAt]
      properties:
        id: { type: integer }
        title: { type: string }
        description: { type: [string, "null"] }
        skillArea: { type: [string, "null"] }
        scheduledOn: { type: string }
        location: { type: [string, "null"] }
        villageId: { type: [integer, "null"] }
        villageName: { type: [string, "null"] }
        capacity: { type: [integer, "null"] }
        status: { type: string }
        registeredCount: { type: integer }
        myRegistrationStatus: { type: [string, "null"] }
        createdAt: { type: string }
    TrainingSessionInput:
      type: object
      required: [title, scheduledOn]
      properties:
        title: { type: string, minLength: 1 }
        description: { type: [string, "null"] }
        skillArea: { type: [string, "null"] }
        scheduledOn: { type: string }
        location: { type: [string, "null"] }
        villageId: { type: [integer, "null"] }
        capacity: { type: [integer, "null"] }
    TrainingRegistration:
      type: object
      required: [id, sessionId, userId, status, createdAt]
      properties:
        id: { type: integer }
        sessionId: { type: integer }
        userId: { type: integer }
        memberName: { type: [string, "null"] }
        membershipCode: { type: [string, "null"] }
        status: { type: string }
        completedAt: { type: [string, "null"] }
        createdAt: { type: string }
    TrainingProgressBody:
      type: object
      required: [userIds, status]
      properties:
        userIds:
          type: array
          items: { type: integer }
          minItems: 1
        status:
          type: string
          enum: [attended, completed]
...